Address code review feedback
- Fix isContinueEnabled to be conditional based on wrap/unwrap flow
- Add user-facing error alerts instead of console-only errors
- Change BridgeFees to use Decimal instead of Double for precision
- Remove duplicate request function in BridgeAPIService
- Add proper decoding error handling
- Use config-based fee calculation for unwrap instead of hardcoded value
- Extract HIGH_BRIDGE_THRESHOLD as named constant
- Improve Ethereum address validation with hex character checking
- Extract shared formatters to reduce duplication and improve performance
- Use static formatters to avoid creating new instances per row

+++ MobileWallet/Common/Models/BridgeModels.swift
@@ -51,9 +51,9 @@ struct BridgeConfig: Codable {
 }
 
 struct BridgeFees {
-    let feeAmount: Double
-    let amountAfterFee: Double
-    let feePercentage: Double
+    let feeAmount: Decimal
+    let amountAfterFee: Decimal
+    let feePercentage: Decimal
     let isOverHighBridgeThreshold: Bool
 }
